# interface/openapi.yaml
openapi: 3.0.3
info:
  title: Freezer Tracking API
  version: 3.0.0
  description: API for freezer tracking application.
servers:
  - url: http://localhost:8080
security:
  - ApiKeyAuth: []
paths:
...
  /accounts:
    get:
      summary: List accounts
      description: Admin only.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: A list of accounts.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create an account
      description: Admin only.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreate"
      responses:
        "201":
          description: Account created.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: Get an account
      description: Admin only.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: Account found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      summary: Update an account
      description: Admin only.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdate"
      responses:
        "200":
          description: Account updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Account"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      summary: Delete an account
      description: Admin only.
      security:
        - ApiKeyAuth: []
      responses:
        "204":
          description: Account deleted.
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}/api-keys:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      summary: List API key sessions for an account
      description: Admin only.
      security:
        - ApiKeyAuth: []
      responses:
        "200":
          description: A list of API key sessions.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountApiKeySession"
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}/api-keys/revoke-all:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Revoke all API key sessions for an account
      description: Admin only.
      security:
        - ApiKeyAuth: []
      responses:
        "204":
          description: Sessions revoked.
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /accounts/{id}/api-keys/revoke:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    post:
      summary: Revoke a specific API key session
      description: Admin only.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RevokeApiKeyRequest"
      responses:
        "204":
          description: Session revoked.
        "403":
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Account not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
    Account:
      type: object
      required:
        - id
        - username
        - isAdmin
        - isSuperAdmin
        - houseId
      properties:
        id:
          type: string
          format: uuid
        username:
          type: string
          format: email
        isAdmin:
          type: boolean
        isSuperAdmin:
          type: boolean
        houseId:
          type: string
          format: uuid
    AccountApiKeySession:
      type: object
      required:
        - apiKey
        - clientType
        - createdAt
        - expiresAt
      properties:
        apiKey:
          type: string
          format: uuid
        clientType:
          $ref: "#/components/schemas/ApiClientType"
        createdAt:
          type: string
          format: date-time
        expiresAt:
          type: string
          format: date-time
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        deviceLabel:
          type: string
          nullable: true
...
    AccountCreate:
      type: object
      required:
        - username
        - password
        - houseId
      properties:
        username:
          type: string
          format: email
        password:
          type: string
          format: password
        isAdmin:
          type: boolean
        isSuperAdmin:
          type: boolean
        houseId:
          type: string
          format: uuid
    AccountUpdate:
      type: object
      required:
        - username
        - houseId
      properties:
        username:
          type: string
          format: email
        isAdmin:
          type: boolean
        isSuperAdmin:
          type: boolean
        houseId:
          type: string
          format: uuid
    RevokeApiKeyRequest:
      type: object
      required:
        - apiKey
      properties:
        apiKey:
          type: string
          format: uuid
...
    ApiClientType:
      type: string
      enum:
        - WEB
        - IOS
...
    ErrorResponse:
      type: object
      required:
        - timestamp
        - status
        - error
        - message
      properties:
        timestamp:
          type: string
          format: date-time
        status:
          type: integer
        error:
          type: string
        message:
          type: string
...